backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
google-generativeai
pydantic-settings
pydantic
httpx